+++ backend/integration_tests/src/fire_and_forget_handler_tests.rs
@@ -1,3 +1,4 @@
+use crate::client::{start_canister, stop_canister};
 use crate::env::ENV;
 use crate::utils::tick_many;
 use crate::{TestEnv, client};
@@ -27,7 +28,7 @@ fn retries_after_failures(failures: usize) {
     assert_eq!(events_response1.events.len(), 1);
     assert!(matches!(events_response1.events[0].event, ChatEvent::Message(_)));
 
-    env.stop_canister(user2.user_id.into(), Some(user2.local_user_index)).unwrap();
+    stop_canister(env, user2.local_user_index, user2.user_id.into());
 
     client::user::happy_path::add_reaction(env, &user1, user2.user_id, "1", message_id);
 
@@ -36,9 +37,7 @@ fn retries_after_failures(failures: usize) {
         env.tick();
     }
 
-    env.start_canister(user2.user_id.into(), Some(user2.local_user_index))
-        .unwrap();
-
+    start_canister(env, user2.local_user_index, user2.user_id.into());
     env.tick();
     env.advance_time(Duration::from_secs(100));
     tick_many(env, 3);