+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Remove `local_group_index_canister_id` from Group/Community init args ([#7917](https://github.com/open-chat-labs/open-chat/pull/7917))
 
+### Fixed
+
+- Fix pushing events to Group/Community canisters ([#7921](https://github.com/open-chat-labs/open-chat/pull/7921))
+
 ## [[2.0.1728](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1728-local_user_index)] - 2025-05-07
 
 ### Added